Create blackjack.py
The deck is unlimited in size. 
There are no jokers. 
The Jack/Queen/King all count as 10.
The Ace can count as 11 or 1.
Use the following list as the deck of cards:
cards = [11, 2, 3, 4, 5, 6, 7, 8, 9, 10, 10, 10, 10]
The cards in the list have an equal probability of being drawn.
Cards are not removed from the deck as they are drawn.